Finish traces and start silkscreen
Still need to modify traces to be more optimal and increase readibility.
Also updated silkscreen preferences locally to have the label of each
switch on both the front and back of the PCB. Need to make a note that
this change is local because of the submodule imported from the start.